Updated map, added collection points
My main goal was to add collection points, but I had to use a different shape file to do it. I don't like the new projection so my options are to figure out if I can change the projection of the current map, or how to translate the GPS coordinates so they fit better with the MN GIS provided map. I also want to change the point colors to match the R color scheme, add a distance bar, add a title, and I should probably also add Wisconsin to the map. Right now, I am committing my changes in case I break something.