Inelegant solution to different colored points
Found an imperfect solution to having each lake be a different colored point, matching the color scheme used in R. Also figured out how to change the plotting character so that _Z. aquatica_ can be a triangle. I also tried to add a scale bar, but the package that I think I need to use (cartopy) is being difficult and I can't install it. Still needed: add Wisconsin and add a legend.