Removed Big Fork River, Phantom Lake, & Mudhen Lake
The code to plot Big Fork River, Phantom Lake, and Mudhen Lake are still in the code... Comments out,  but for now I don't want to keep them. I can't get the projection quite right with both Minnesota & Wisconsin so that is why I have removed Mudhen and Phantom Lakes. We probably won't include Big Fork River in the publication, so that's why I removed that one.